+++ czkawka_core/src/same_music.rs
@@ -23,10 +23,8 @@ use symphonia::core::io::MediaSourceStream;
 use symphonia::core::meta::MetadataOptions;
 use symphonia::core::probe::Hint;
 
-use crate::common::{
-    create_crash_message, filter_reference_folders_generic, load_cache_from_file_generalized, prepare_thread_handler_common, save_cache_to_file_generalized,
-    send_info_and_wait_for_ending_all_threads, AUDIO_FILES_EXTENSIONS,
-};
+use crate::common::{create_crash_message, filter_reference_folders_generic, prepare_thread_handler_common, send_info_and_wait_for_ending_all_threads, AUDIO_FILES_EXTENSIONS};
+use crate::common_cache::{get_similar_music_cache_file, load_cache_from_file_generalized, save_cache_to_file_generalized};
 use crate::common_dir_traversal::{CheckingMethod, DirTraversalBuilder, DirTraversalResult, FileEntry, ProgressData, ToolType};
 use crate::common_tool::{CommonData, CommonToolData};
 use crate::common_traits::*;
@@ -235,7 +233,8 @@ impl SameMusic {
         let mut non_cached_files_to_check: BTreeMap<String, MusicEntry> = Default::default();
 
         if self.common_data.use_cache {
-            let (messages, loaded_items) = load_cache_from_file_generalized::<MusicEntry>(get_cache_file(checking_tags), self.get_delete_outdated_cache(), &self.music_to_check);
+            let (messages, loaded_items) =
+                load_cache_from_file_generalized::<MusicEntry>(get_similar_music_cache_file(checking_tags), self.get_delete_outdated_cache(), &self.music_to_check);
             self.get_text_messages_mut().extend_with_another_messages(messages);
             loaded_hash_map = loaded_items.unwrap_or_default();
 
@@ -266,7 +265,7 @@ impl SameMusic {
             all_results.insert(file_entry.path.to_string_lossy().to_string(), file_entry);
         }
 
-        let messages = save_cache_to_file_generalized(get_cache_file(checking_tags), &all_results, self.common_data.save_also_as_json);
+        let messages = save_cache_to_file_generalized(get_similar_music_cache_file(checking_tags), &all_results, self.common_data.save_also_as_json);
         self.get_text_messages_mut().extend_with_another_messages(messages);
         debug!("save_cache - end");
     }
@@ -905,15 +904,6 @@ fn read_single_file_tag(path: &str, music_entry: &mut MusicEntry) -> bool {
     music_entry.bitrate = bitrate;
 
     true
-}
-
-// Using different cache folders, because loading cache just for finding duplicated tags would be really slow
-fn get_cache_file(checking_tags: bool) -> &'static str {
-    if checking_tags {
-        "cache_same_music_tags_61.bin"
-    } else {
-        "cache_same_music_fingerprints_61.bin"
-    }
 }
 
 impl Default for SameMusic {
